Change bot name from Duke to Boti
