+++ src/main/java/boti/exception/InvalidCommandException.java
@@ -1,0 +1,13 @@
+package boti.exception;
+
+/**
+ * The Exception class for invalid command
+ */
+public class InvalidCommandException extends Exception {
+    /**
+     * The constructor of InvalidCommandException
+     */
+    public InvalidCommandException() {
+        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
+    }
+}
