+++ src/main/java/boti/exception/InvalidDeadlineException.java
@@ -1,0 +1,13 @@
+package boti.exception;
+
+/**
+ * The exception for invalid deadline command
+ */
+public class InvalidDeadlineException extends Exception {
+    /**
+     * The constructor of InvalidDeadlineException
+     */
+    public InvalidDeadlineException() {
+        super("☹ OOPS!!! I'm sorry, but the Deadline input is invalid :-(");
+    }
+}
