--- src/main/java/duke/exception/InvalidCommandException.java
@@ -1,13 +1,0 @@
-package duke.exception;
-
-/**
- * The Exception class for invalid command
- */
-public class InvalidCommandException extends Exception {
-    /**
-     * The constructor of InvalidCommandException
-     */
-    public InvalidCommandException() {
-        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
-    }
-}
