--- src/main/java/duke/exception/InvalidEventException.java
@@ -1,13 +1,0 @@
-package duke.exception;
-
-/**
- * The Exception for invalid event command
- */
-public class InvalidEventException extends Exception {
-    /**
-     * The constructor of InvalidEventException
-     */
-    public InvalidEventException() {
-        super("☹ OOPS!!! I'm sorry, but the input Event is invalid :-(");
-    }
-}
