+++ src/test/java/boti/task/ToDoSample.java
@@ -1,0 +1,10 @@
+package boti.task;
+
+/**
+ * Sample for ToDo
+ */
+public class ToDoSample extends ToDo {
+    public ToDoSample() {
+        super("TaskSample");
+    }
+}
