--- src/test/java/duke/task/ToDoSample.java
@@ -1,10 +1,0 @@
-package duke.task;
-
-/**
- * Sample for ToDo
- */
-public class ToDoSample extends ToDo {
-    public ToDoSample() {
-        super("TaskSample");
-    }
-}
