+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-    Hello! I'm Not A ChatBot
+    Hello! I'm Not A Boti
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
